requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
python-dotenv==1.0.0
schedule==1.2.0
newspaper3k==0.2.8
nltk==3.8.1
jieba==0.42.1
pyyaml==6.0.1
jinja2==3.1.2
yfinance==0.2.28
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
aiofiles==23.2.1 